refactor: move validate_ios_relay_binding method into http_client
