+++ camera_hub/src/notification_target.rs
@@ -2,14 +2,12 @@
 //!
 //! SPDX-License-Identifier: GPL-3.0-or-later
 
-use secluso_client_lib::http_client::{HttpClient, IosRelayBinding, NotificationTarget};
+use secluso_client_lib::http_client::{validate_ios_relay_binding, HttpClient, NotificationTarget};
 use std::fs;
 use std::io;
 use std::path::Path;
-use url::Url;
 
 const TARGET_FILENAME: &str = "notification_target.json";
-const TRUSTED_IOS_RELAY_HOSTS: &[&str] = &["relay.secluso.com", "testing-relay.secluso.com"];
 
 // Build the placeholder target we keep for iOS while no relay binding is available / after the current binding has been rejected.
 fn ios_placeholder_target(platform: &str) -> NotificationTarget {
@@ -20,80 +18,6 @@ fn ios_placeholder_target(platform: &str) -> NotificationTarget {
         unifiedpush_pub_key: None,
         unifiedpush_auth: None,
     }
-}
-
-// Mirror the server-side relay checks before the hub sends any outbound iOS request.
-// Ensures a malicious/stale notification target cannot turn the hub into a generic HTTPS client.
-fn validate_ios_relay_base_url(raw_url: &str) -> io::Result<Url> {
-    let parsed = Url::parse(raw_url)
-        .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e.to_string()))?;
-    if parsed.scheme() != "https" {
-        return Err(io::Error::new(
-            io::ErrorKind::InvalidInput,
-            "iOS relay base URL must use https",
-        ));
-    }
-    if !parsed.username().is_empty() || parsed.password().is_some() {
-        return Err(io::Error::new(
-            io::ErrorKind::InvalidInput,
-            "iOS relay base URL must not include credentials",
-        ));
-    }
-    if parsed.query().is_some() || parsed.fragment().is_some() {
-        return Err(io::Error::new(
-            io::ErrorKind::InvalidInput,
-            "iOS relay base URL must not include a query or fragment",
-        ));
-    }
-    if parsed.path() != "/" {
-        return Err(io::Error::new(
-            io::ErrorKind::InvalidInput,
-            "iOS relay base URL must not include a path prefix",
-        ));
-    }
-
-    let host = parsed.host_str().ok_or_else(|| {
-        io::Error::new(
-            io::ErrorKind::InvalidInput,
-            "iOS relay base URL is missing a host",
-        )
-    })?;
-    let port = parsed.port_or_known_default().ok_or_else(|| {
-        io::Error::new(
-            io::ErrorKind::InvalidInput,
-            "iOS relay base URL is missing an https port",
-        )
-    })?;
-    if port != 443 {
-        return Err(io::Error::new(
-            io::ErrorKind::InvalidInput,
-            "iOS relay base URL must use the default https port",
-        ));
-    }
-    if !TRUSTED_IOS_RELAY_HOSTS
-        .iter()
-        .any(|allowed| host.eq_ignore_ascii_case(allowed))
-    {
-        return Err(io::Error::new(
-            io::ErrorKind::InvalidInput,
-            format!("Refusing unexpected iOS relay host: {host}"),
-        ));
-    }
-
-    Ok(parsed)
-}
-
-fn validate_ios_relay_binding(binding: &IosRelayBinding) -> io::Result<()> {
-    let relay_base = binding.relay_base_url.trim();
-    if relay_base.is_empty() {
-        return Err(io::Error::new(
-            io::ErrorKind::InvalidInput,
-            "iOS relay base URL is required",
-        ));
-    }
-    validate_ios_relay_base_url(relay_base)?;
-
-    Ok(())
 }
 
 pub fn persist_notification_target(state_dir: &str, target: &NotificationTarget) -> io::Result<()> {
@@ -205,49 +129,4 @@ pub fn send_notification(
     }
 
     http_client.send_fcm_notification(notification_msg)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{validate_ios_relay_base_url, validate_ios_relay_binding};
-    use secluso_client_lib::http_client::IosRelayBinding;
-
-    // Build an otherwise-valid relay binding and let each test vary only the relay base URL it wants to validate.
-    fn ios_binding(relay_base_url: &str) -> IosRelayBinding {
-        IosRelayBinding {
-            relay_base_url: relay_base_url.to_string(),
-            hub_token: "hub-token".to_string(),
-            app_install_id: "install-id".to_string(),
-            hub_id: "hub-id".to_string(),
-            device_token: "device-token".to_string(),
-            expires_at_epoch_ms: 1,
-        }
-    }
-
-    #[test]
-    // Tests that the camera hub accepts the public production relay.
-    fn accepts_trusted_ios_relay_host() {
-        validate_ios_relay_base_url("https://relay.secluso.com")
-            .expect("trusted relay host should be accepted");
-    }
-
-    #[test]
-    // Tests that server-side iOS relay checks reject unexpected relay hosts before the target can be persisted to the hub.
-    fn rejects_untrusted_ios_relay_host() {
-        let err = validate_ios_relay_base_url("https://evil.example")
-            .expect_err("unexpected relay host should be rejected");
-
-        assert!(err
-            .to_string()
-            .contains("Refusing unexpected iOS relay host"));
-    }
-
-    #[test]
-    // Tests that the binding-level check rejects incomplete relay bindings before send_notification hands them to the HTTP client.
-    fn rejects_empty_ios_relay_base_url() {
-        let err = validate_ios_relay_binding(&ios_binding("   "))
-            .expect_err("empty relay base URL should be rejected");
-
-        assert!(err.to_string().contains("iOS relay base URL is required"));
-    }
 }
